// apps/manager/prisma/generated/zod/inputTypeSchemas/UserMcpServerInstanceToolGroupMaxOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const UserMcpServerInstanceToolGroupMaxOrderByAggregateInputSchema: z.ZodType<Prisma.UserMcpServerInstanceToolGroupMaxOrderByAggregateInput> = z.object({
  mcpServerInstanceId: z.lazy(() => SortOrderSchema).optional(),
  toolGroupId: z.lazy(() => SortOrderSchema).optional(),
  sortOrder: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default UserMcpServerInstanceToolGroupMaxOrderByAggregateInputSchema;