// apps/manager/prisma/generated/zod/inputTypeSchemas/UserMcpServerInstanceToolGroupMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const UserMcpServerInstanceToolGroupMinOrderByAggregateInputSchema: z.ZodType<Prisma.UserMcpServerInstanceToolGroupMinOrderByAggregateInput> = z.object({
  mcpServerInstanceId: z.lazy(() => SortOrderSchema).optional(),
  toolGroupId: z.lazy(() => SortOrderSchema).optional(),
  sortOrder: z.lazy(() => SortOrderSchema).optional(),
  createdAt: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default UserMcpServerInstanceToolGroupMinOrderByAggregateInputSchema;